requirements.txt
python-telegram-bot==21.0.1
requests==2.31.0
aiohttp==3.9.1
python-dotenv==1.0.0